Improve UX of pinning feature
Students no longer see a tooltip nor any button affordance (i.e. cursor
and color change on hover/focus) on the pinning feature. For moderators,
the tooltip is now correct when a thread is pinned, the button is
accessible for keyboard-only users and screen readers, and an error
dialog now appears if an error occurs in attempting to pin/unpin.

JIRA: FOR-192